# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Swagger Persons",
        "license": {
            "name": "MIT"
        }
    },
    "host": "persons.swagger.io",
    "basePath": "/v1",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/persons": {
            "get": {
                "summary": "List all persons",
                "operationId": "listpersons",
                "tags": [
                    "persons"
                ],
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "How many items to return at one time (max 100)",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An paged array of persons",
                        "headers": {
                            "x-next": {
                                "type": "string",
                                "description": "A link to the next page of responses"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/persons"
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a person",
                "operationId": "createPerson",
                "tags": [
                    "persons"
                ],
                "parameters": [
                    {
                      "name": "name",
                      "in": "body",
                      "description": "Name to add to the store",
                      "required": true,
                      "schema": {
                        "$ref": "#/definitions/NewPerson"
                      }
                    }
                  ],
                "responses": {
                    "201": {
                        "description": "Create Person"
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/persons/{index}": {
            "get": {
                "summary": "Info for a specific person",
                "operationId": "showPersonById",
                "tags": [
                    "persons"
                ],
                "parameters": [
                    {
                        "name": "index",
                        "in": "path",
                        "required": true,
                        "description": "The id of the pet to retrieve",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expected response to a valid request",
                        "schema": {
                            "$ref": "#/definitions/persons"
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update for a specific person",
                "operationId": "updateById",
                "tags": [
                    "persons"
                ],
                "parameters": [
                    {
                        "name": "index",
                        "in": "path",
                        "required": true,
                        "description": "The id of the pet to retrieve",
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "body",
                        "description": "Name to add to the store",
                        "required": true,
                        "schema": {
                          "$ref": "#/definitions/NewPerson"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expected response to a valid request",
                        "schema": {
                            "$ref": "#/definitions/persons"
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete for a specific person",
                "operationId": "deletePersonById",
                "tags": [
                    "persons"
                ],
                "parameters": [
                    {
                        "name": "index",
                        "in": "path",
                        "required": true,
                        "description": "The id of the pet to retrieve",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expected response to a valid request",
                        "schema": {
                            "$ref": "#/definitions/persons"
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Persons": {
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                }
            }
        },
        "persons": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/Pet"
            }
        },
        "Error": {
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "NewPerson": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "tag": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }